refactor MIME parser
