+++ src/app/modules/email/classes/mime-message-header/mime-message-header.spec.ts
@@ -1,0 +1,7 @@
+import { MimeMessageHeader } from "./mime-message-header";
+
+describe('MimeMessageHeader', () => {
+  it('should create an instance', () => {
+    expect(new MimeMessageHeader()).toBeTruthy();
+  });
+});
