+++ src/app/modules/email/classes/mime-message-part/mime-message-part.spec.ts
@@ -1,0 +1,7 @@
+import { MimeMessagePart } from "./mime-message-part";
+
+describe('MimeMessagePart', () => {
+  it('should create an instance', () => {
+    expect(new MimeMessagePart()).toBeTruthy();
+  });
+});
