+++ src/app/modules/email/classes/mime-message/mime-message.spec.ts
@@ -1,0 +1,7 @@
+import { MimeMessage } from "./mime-message";
+
+describe('MimeMessage', () => {
+  it('should create an instance', () => {
+    expect(new MimeMessage()).toBeTruthy();
+  });
+});
